% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BuildCascade.R
\name{BuildCascade}
\alias{BuildCascade}
\title{Build a Cascade Model Network}
\usage{
BuildCascade(S = 50, C = 0.1)
}
\arguments{
\item{S}{Number of species in network; defaults to 50}

\item{C}{Network connectance, defined as \eqn{2L/(S(S-1))},
  where L is the number of links. Defaults to .1. The resulting
  matrix will have a connectance of C +/- \eqn{2sqrt(.5C(1-C)S(S-1))}.}
}
\value{
A weakly connected, unsigned adjacency matrix of size (S,S).
}
\description{
Generates a random network of a given size and approximate
connectance according to the cascade model
}
\examples{
BuildCascade()
}
\references{
Cohen, J.E. and C.M. Newman. 1985. A Stochastic Theory of
  Community Food Webs: I. Models and Aggregated Data. Proc. R. Soc.
  Lond. B. 224: 421-448; DOI: 10.1098/rspb.1985.0042.
}

