% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildMPN.R
\name{BuildMPN}
\alias{BuildMPN}
\title{Build a Minimum Potential Niche Model Network}
\usage{
BuildMPN(S = 50, C = 0.1, GapProb = 0.25)
}
\arguments{
\item{S}{Number of species in network; defaults to 50}

\item{C}{Network connectance, defined as 2L/(S(S-1)),
where L is the number of links. Defaults to .1. The resulting
matrix will have a connectance of C +/- 2sqrt(.5C(1-C)S(S-1)).}

\item{GapProb}{Probability of a gap in a feeding niche; defaults to .25.}
}
\value{
M A weakly connected, unsigned adjacency matrix of size (S,S).
}
\description{
Generates a random network of a given size and approximate
connectance according to the minimum potential niche model.
}
\examples{
BuildMPN()

}
\references{
Allesina, S., D. Alonso, and M. Pascual. 2008.
  A General Model for Food Web Structure. Science 320(5876): 658-661.
}

