% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/BuildNiche.R
\name{BuildNiche}
\alias{BuildNiche}
\title{Build a Niche Model Network}
\usage{
BuildNiche(S = 50, C = 0.1)
}
\arguments{
\item{S}{Number of species in network; defaults to 50}

\item{C}{Network connectance, defined as 2L/(S(S-1)),
  where L is the number of links. Defaults to .1. The resulting
  matrix will have a connectance of C +/- 2sqrt(.5C(1-C)S(S-1)).}
}
\value{
M A weakly connected, unsigned adjacency matrix of size (S,S).
}
\description{
Generates a random network of a given size and approximate
connectance according to the niche model
}
\examples{
BuildNiche()
}
\references{
Williams, R.J., and N.D. Martinez. 2000.
  Simple rules yield complex food webs. Nature 404(6774): 180-3.
}

