% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{caricaie}
\alias{caricaie}
\title{The food web of the marsh at Grande Caricaie at Lake Neuch\^{a}tel,
Switzerland.}
\format{An adjacency matrix with 163 rows and 163 columns.}
\description{
The food web of the marsh at Grande Caricaie at Lake Neuch\^{a}tel,
Switzerland.
}
\references{
Cattin Blandenier, M.-F. Food web ecology: models and
application to conservation. PhD thesis, (2004).
}

