% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discreteLV_R.R
\name{discreteLV}
\alias{discreteLV}
\title{Simulate a discrete-time Lotka-Volterra time series}
\usage{
discreteLV(rmat, alphas, n0s, Is, deltat, simtime)
}
\arguments{
\item{rmat}{a matrix with S (number of species) rows and simtime
cols containing the growth rates at each time step (potentially
perturbed at each time step)}

\item{alphas}{an S by S matrix of per capita interaction strengths}

\item{n0s}{a vector of length S containing species abundances}

\item{Is}{immigration rates}

\item{deltat}{the size of the time step}

\item{simtime}{the total number of time steps to simulate}
}
\value{
an S by simtime matrix of species abundances at each timestep
}
\description{
Simulate discrete-time Lotka-Volterra population dynamics
with perturbed growth rates
}

