% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{discreteLV_C}
\alias{discreteLV_C}
\title{Simulate a discrete-time Lotka-Volterra time series in C}
\usage{
discreteLV_C(r_, alphas_, ntmp_, deltat, simtime)
}
\arguments{
\item{deltat}{the size of the time step}

\item{simtime}{the total number of time steps to simulate}

\item{rmat}{a matrix with S (number of species) rows and simtime
cols containing the growth rates at each time step (potentially
perturbed at each time step)}

\item{alphas}{an S by S matrix of per capita interaction strengths}

\item{n0s}{a vector of length S containing species abundances}
}
\value{
an S by simtime matrix of species abundances at each timestep
}
\description{
Simulate discrete-time Lotka-Volterra population dynamics
with perturbed growth rates
}

