% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{flensburg}
\alias{flensburg}
\title{The food web of Flensburg Fjord in the Baltic Sea, between Germany and Denmark.}
\format{An adjacency matrix with 180 rows and 180 columns.}
\description{
The food web of Flensburg Fjord in the Baltic Sea, between Germany and Denmark.
}
\references{
Zander, C. D., Josten, N., Detloff, K. C., Poulin, R.,
McLaughlin, J. P., and Thieltges, D. W. Food web including metazoan
parasites for a brackish shallow water ecosystem in Germany and Denmark:
Ecological Archives E092-174. Ecology 92(10), 2007-2007 (2011).
}

