% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetCentralities.R
\name{GetCentralities}
\alias{GetCentralities}
\title{Calculate network metrics for a food web}
\usage{
GetCentralities(matfile)
}
\arguments{
\item{matfile}{path to parameterized nework}
}
\value{
a data frame containing log(closeness centrality),
log(eigenvector centrality), and trophic level for each
species in the network.
}
\description{
Read in a matrix and calculate log closeness centrality,
log eigenvector centrality, and trophic level for each species.
Note that this function is used internally and is not likely to be
useful on its own. Use igraph functionality and getTL() for centrality
calculations beyond replicating simulations from Wootton et al.
}

