% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTL.R
\name{getTL}
\alias{getTL}
\title{Calculate prey-averaged trophic levels}
\usage{
getTL(M)
}
\arguments{
\item{M}{An S by S adjacency matrix, where S is the number of species}
}
\value{
A vector of S trophic levels
}
\description{
Calculate prey-averaged trophic levels for each species in a matrix
}
\examples{
mat <- BuildCascade()
getTL(mat)
}
\references{
Williams, R.J. and N.D. Martinez. 2004. Limits
  to Trophic Levels and Omnivory in Complex Food Webs:
  Theory and Data. Am. Nat.
}

