% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{hasTrailingSlash}
\alias{hasTrailingSlash}
\title{Utility functions for handling path and file names}
\usage{
hasTrailingSlash(dirpath)
}
\arguments{
\item{dirpath}{a string containing a path to a directory}

\item{filepath}{a string containing a path to a file}
}
\value{
hasTrailiingSlash returns a boolean flag. addTrailingSlash
returns the directory path, and stripFileExtension returns the file
with the extension removed.
}
\description{
hasTrailingSlash checks for a trailing slash '/' in a path string,
addTrailingSlash adds a slash if it isn't already present, and
stripFileExtension strips the final file extension (.txt, etc.)
from a given file name
}

