% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadEmpiricalScript.R
\name{loadEmpiricalScript}
\alias{loadEmpiricalScript}
\title{Parameterize empirical networks, run models, and plot results as in Wootton et al.}
\usage{
loadEmpiricalScript()
}
\description{
Parameterizes empirical networks 30 times, simulates
population dynamics, runs hierarchical model, and generates violin plots
for the following food webs: caricaie, flensburg, otago, reef, serengeti,
stmarks, sylt, tatoosh, and ythan. This replicates main text results
in Wootton et al. (in prep). Note that this simulation and analysis will likely
days to complete!
}

