% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/LognormalParam.R
\name{LognormalParam}
\alias{LognormalParam}
\title{Parameterize a Lotka-Volterra system given an adjacency matrix}
\usage{
LognormalParam(Adj)
}
\arguments{
\item{Adj}{An adjacency matrix}
}
\value{
a list with elements Mat (a matrix of interaction strengths) and
  Pop (a vector of equilibrium population sizes).
}
\description{
Parameterize an adjacency matrix with interaction strengths and
equilibrium abundances. Interaction strengths and abundances are drawn from a
lognormal distribution truncated 5 standard deviations above the
mean.
}
\examples{
mat <- BuildCascade()
LognormalParam(mat)
}

