% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MeanSigmaResiduals.R
\name{MeanSigmaResiduals}
\alias{MeanSigmaResiduals}
\title{Regress log(Sigma) on log(Mean) and Calculate Residuals}
\usage{
MeanSigmaResiduals(Mean, Sigma)
}
\arguments{
\item{Mean}{A vector of mean population sizes for a single simulation
of a single parameterization of a single network structure}

\item{Sigma}{An associated vector of standard deviations of population
sizes for a single simulation of a single parameterization of a single
network structure.}
}
\value{
if the regression is significant, a vector of residual values. If the
log(Sigma) ~ log(Mean) linear relationship is not significant, returns NULL.
}
\description{
Run a linear regression to find the relationship between log mean and log
standard deviation for one simulation of a parameterization of a network structure.
The residuals can then be used as a measure of variability; that is,
deviation from expected variability under Taylor's Law.
}

