% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/normalizeM.R
\name{normalizeM}
\alias{normalizeM}
\title{Normalize matrix columns}
\usage{
normalizeM(M)
}
\arguments{
\item{M}{a matrix to be normalized}
}
\value{
the matrix M with columns normalized by column sum
}
\description{
Normalize the columns of the ajacency matrix M
by its column sums. This is a utility function for
predictimportance which will likely not be useful
except internally.
}

