% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{otago}
\alias{otago}
\title{The food web of the intertidal mudflat in Otago Harbour, New Zealand.}
\format{An adjacency matrix with 180 rows and 180 columns.}
\description{
The food web of the intertidal mudflat in Otago Harbour, New Zealand.
}
\references{
Mouritsen, K. N., Poulin, R., McLaughlin, J. P., and
Thieltges, D. W. Food web including metazoan parasites for an intertidal
ecosystem in New Zealand: Ecological Archives E092-713. Ecology 92(10),
2006-2006 (2011).
}

