% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\docType{package}
\name{predictimportance}
\alias{predictimportance}
\alias{predictimportance-package}
\title{predictimportance: a package containing code and data necessary to
replicate results in Wootton et al. (in prep.).}
\description{
The predictimportance package contains functions to randomly parameterize
empirical networks, generate and parameterize model food webs, simulate
community dynamics, and predict species importance using several covariates
using a hierarchical modelling approach. The highest-level and most
user-friendly functions are runScriptsEmpirical() and runScriptsModel(), which
can be used to replicate all or part of the data simulation and analysis in
a single function call. For users who wish to replicate all results, the
simplest approach is to run the scripts in inst/scripts, FullSimulationEmpirical.R
(for main text results) and FullSimulationModels.R (for supplemental results).
These scripts run perform all data simulation and analysis, and plot violin plots
to summarize the results. For users who wish to analyze their own empirical
dataset, it is recommended to use the Step1_Empirical_Parameterization(),
Step2_Discrete_LV(), and Step3_Hierarchical_Model() functions in order.
Note that it is worthwhile to examine output from Step 2 before determining
if the specific hierarchical model in Step 3 is appropriate for your data.
Note that Step 2 (simulation of community dynamics) may produce non-finite
values for large networks, due to precision issues.
}
\references{
Wootton, J.T., Sander, E.L., Barab\'as, Gy\"orgy, and A. Henry.
In prep.
}

