% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{reef}
\alias{reef}
\title{The food web for the Caribbean reef on the American Virgin Islands.}
\format{An adjacency matrix with 249 rows and 249 columns.}
\description{
The food web for the Caribbean reef on the American Virgin Islands.
}
\references{
Opitz, S. Trophic interactions in Caribbean coral reefs.
Number 1085. WorldFish, (1996).
}

