% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{serengeti}
\alias{serengeti}
\title{The terrestrial food web of the Serengeti grassland in Tanzania.}
\format{An adjacency matrix with 170 rows and 170 columns.}
\description{
The terrestrial food web of the Serengeti grassland in Tanzania.
}
\references{
Baskerville, E. B., Dobson, A. P., Bedford, T., Allesina, S.,
Anderson, T. M., and Pascual, M. Spatial guilds in the Serengeti food web
revealed by a Bayesian group model. PLoS Computational Biology 7(12),
e1002321 (2011).
}

