% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Step1_Empirical.R
\name{Step1_Empirical_Parameterization}
\alias{Step1_Empirical_Parameterization}
\title{Step 1: Generate and/or parameterize Lotka-Volterra communities}
\usage{
Step1_Empirical_Parameterization(web, foldname = web, path = "Data",
  nruns = 30, Exponential = FALSE, seed = NULL)
}
\arguments{
\item{web}{empirical network
name ("caricaie", "otago", "serengeti", "sylt", "ythan", "flensburg",
"reef", "stmarks", or "tatoosh")}

\item{foldname}{Folder name where matrices and abundance vectors should
be written}

\item{path}{Path where data are kept. 'foldname' will be created as a folder
here if it does not already exist}

\item{nruns}{Number of parameterizations to be generated.}

\item{Exponential}{TEMPORARY}

\item{seed}{Random seed for reproducibility}
}
\value{
This function does not return an object, but it writes parameterized
  matrices (with file names ending in '-mat.txt') and vectors of equilibrium
  abundances (with file names ending in '-pop.txt') to the folder
  path/foldname.
}
\description{
Generates parameterized networks and abundance data that can be
simulated in Step 2.
}

