% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Step1.R, R/Step1_Model.R
\name{Step1_Generate_Networks}
\alias{Step1_Generate_Networks}
\title{Step 1: Generate and/or parameterize Lotka-Volterra communities}
\usage{
Step1_Generate_Networks(web = "Cascade", foldname = web, path = "Data",
  S = 50, C = 0.1, GapProb = 0.25, nwebs = 30, nruns = 30,
  seed = NULL)

Step1_Generate_Networks(web = "Cascade", foldname = web, path = "Data",
  S = 50, C = 0.1, GapProb = 0.25, nwebs = 30, nruns = 30,
  seed = NULL)
}
\arguments{
\item{web}{Model name ('Cascade', 'Niche', or 'MPN') or empirical network
name ("caricaie", "otago", "serengeti", "sylt", "ythan", "flensburg",
"reef", "stmarks", or "tatoosh")}

\item{foldname}{Folder name where matrices and abundance vectors should
be written}

\item{path}{Path where data are kept. 'foldname' will be created as a folder
here if it does not already exist}

\item{S}{Number of species in the generated networks}

\item{C}{Network connectance, defined as 2L/(S(S-1)), where L
is the number of links.}

\item{GapProb}{Probability of a gap in a niche. Only used if web == 'MPN'.}

\item{nwebs}{Number of random web structures to be generated.}

\item{nruns}{Number of parameterizations to be generated.}

\item{seed}{Random seed for reproducibility}

\item{web}{Model name ('Cascade', 'Niche', or 'MPN') or empirical network
name ("caricaie", "otago", "serengeti", "sylt", "ythan", "flensburg",
"reef", "stmarks", or "tatoosh")}

\item{foldname}{Folder name where matrices and abundance vectors should
be written}

\item{path}{Path where data are kept. 'foldname' will be created as a folder
here if it does not already exist}

\item{S}{Number of species in the generated networks}

\item{C}{Network connectance, defined as 2L/(S(S-1)), where L
is the number of links.}

\item{GapProb}{Probability of a gap in a niche. Only used if web == 'MPN'.}

\item{nwebs}{Number of random web structures to be generated.}

\item{nruns}{Number of parameterizations to be generated.}

\item{seed}{Random seed for reproducibility}
}
\value{
This function does not return an object, but it writes parameterized
  matrices (with file names ending in '-mat.txt') and vectors of equilibrium
  abundances (with file names ending in '-pop.txt') to the folder
  path/foldname.

This function does not return an object, but it writes parameterized
  matrices (with file names ending in '-mat.txt') and vectors of equilibrium
  abundances (with file names ending in '-pop.txt') to the folder
  path/foldname.
}
\description{
Generates parameterized networks and abundance data that can be
simulated in Step 2.

Generates parameterized networks and abundance data that can be
simulated in Step 2.
}

