% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step2.R
\name{Step2_Discrete_LV}
\alias{Step2_Discrete_LV}
\title{Step 2: Simulate dynamics using discrete-time Lotka-Volterra}
\usage{
Step2_Discrete_LV(path, seed = NULL)
}
\arguments{
\item{path}{path to folder where '-mat.txt' and '-pop.txt' files were
generated in Step 1.}

\item{seed}{random seed, for reproducibility}
}
\value{
Returns the csv file name where results are stored. This file name
  can then be passed to Step 3. This results file is appended to file
  as the simulations are completed.
}
\description{
Takes network parameterizations created in Step 1 and simulates community
dynamics using a discrete-time Lotka-Volterra. Transforms variables for
use in hierarchical model (Step 3).
}

