% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Step3.R
\name{Step3_Hierarchical_Model}
\alias{Step3_Hierarchical_Model}
\title{Step 3: Analyze data using a hierarchical model}
\usage{
Step3_Hierarchical_Model(datafname, PathToResults = "Results",
  empirical = TRUE)
}
\arguments{
\item{datafname}{path to file generated in Step 2 (file name will end
in '-AllData.csv')}

\item{PathToResults}{path to location where results should be stored.}

\item{empirical}{flags if the simulated data are based on an empirical
network structure. This will affect the hierarchical model structure}
}
\value{
This function does not return an object, but it writes hierarchical
  model output to text files, and saves the entire lmer object to a .RData file
}
\description{
Takes data file created in Step 2 and analyzes the data using a hierarchical model
}

