% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{stmarks}
\alias{stmarks}
\title{The food web for the seagrass community at Goose Creek By in St. Marks
National Wildlife Refuge, Florida.}
\format{An adjacency matrix with 143 rows and 143 columns.}
\description{
The food web for the seagrass community at Goose Creek By in St. Marks
National Wildlife Refuge, Florida.
}
\references{
Christian, R. R. and Luczkovich, J. J. Organizing and
understanding a winter's sea-grass foodweb network through effective
trophic levels. Ecological Modelling 117(1), 99-124 (1999).
}

