% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{sylt}
\alias{sylt}
\title{The food web for the North Sea Sylt tidal basin between Germany and
Denmark.}
\format{An adjacency matrix with 230 rows and 230 columns.}
\description{
The food web for the North Sea Sylt tidal basin between Germany and
Denmark.
}
\references{
Thieltges, D. W., Reise, K., Mouritsen, K. N., McLaughlin,
J. P., and Poulin, R. Food web including metazoan parasites for a tidal
basin in Germany and Denmark: Ecological Archives E092-172. Ecology 92(10),
2005-2005 (2011).
}

