% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{tatoosh}
\alias{tatoosh}
\title{Food web for the mussel bed at the Tatoosh intertidal in Tatoosh Island,
Washington.}
\format{An adjacency matrix with 110 rows and 110 columns.}
\source{
\url{http://datadryad.org/review?doi=doi:10.5061/dryad.39jv1}
}
\description{
Food web for the mussel bed at the Tatoosh intertidal in Tatoosh Island,
Washington.
}

