% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/utilities.R
\name{utilities}
\alias{hasTrailingSlash}
\alias{stripFileExtension}
\alias{utilities}
\title{Utility functions for handling path and file names

Check for a trailing slash '/' in a path string and strip
the final file extension (.txt, etc.) from a given file name}
\usage{
hasTrailingSlash(mystring)

stripFileExtension(filepath)
}
\arguments{
\item{mystring}{a character vector}

\item{filepath}{a character vector containing a file name}
}
\value{
hasTrailingSlash() returns 1 if mystring ends in '/', 0 otherwise.
stripFileExtension() returns the filepath character vector with the
final file extension removed.
}
\description{
Utility functions for handling path and file names

Check for a trailing slash '/' in a path string and strip
the final file extension (.txt, etc.) from a given file name
}

