% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datadocumentation.R
\name{ythan}
\alias{ythan}
\title{The food web for Ythan Estuary in Scotland.}
\format{An adjacency matrix with 92 rows and 92 columns.}
\description{
The food web for Ythan Estuary in Scotland.
}
\references{
Cohen, J. E., Schittler, D. N., Raffaelli, D. G., and Reuman, D. C.
Food webs are more than the sum of their tritrophic parts. Proceedings of the
National Academy of Sciences 106(52), 22335-22340 (2009).
}

