% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/runScripts.R
\name{runScripts}
\alias{runScripts}
\alias{runScriptsEmpirical}
\alias{runScriptsModel}
\title{Run all analyses for a food web model or network

Run all data simulation and analysis steps for a given model or
empirical web(s). This may take a while!}
\usage{
runScriptsModel(model = "Cascade", path = "./")

runScriptsEmpirical(web = "All", path = "./")
}
\arguments{
\item{model}{food web model used to construct the network: options are
  'Cascade', 'Niche', 'MPN25', 'MPN35', and 'MPN45'.}

\item{path}{path to where Data and Results folders are or should be created}

\item{web}{empirical network to be parameterized and simulated: options are
  "caricaie", "otago", "serengeti", "sylt", "ythan", "flensburg",
  "reef", "stmarks", and "tatoosh". A vector of network names can be
  provided to simulate more than one network, or web = "All" can be used
  to simulate all networks in the package.}
}
\description{
Run all analyses for a food web model or network

Run all data simulation and analysis steps for a given model or
empirical web(s). This may take a while!
}
\examples{
\dontrun{
runScriptsModel('Niche')
}
\dontrun{
runScriptsEmpirical('tatoosh')
}
}

